requirements.txt
pandas
sqlalchemy
python-dotenv
loguru
psycopg2-binary
